change entities fileds

+++ Benchmark X/bookstore/src/main/java/api/ItemApi.java
@@ -12,7 +12,7 @@ public class ItemApi {
 	private String publisher;
 	private Date publicationDate;
 	private BigDecimal cost;
-	private Long availability;
+	private Integer availability;
 	
 	public Long getId() {
 		return id;
@@ -56,10 +56,10 @@ public class ItemApi {
 	public void setCost(BigDecimal cost) {
 		this.cost = cost;
 	}
-	public Long getAvailability() {
+	public Integer getAvailability() {
 		return availability;
 	}
-	public void setAvailability(Long availability) {
+	public void setAvailability(Integer availability) {
 		this.availability = availability;
 	}
 }
